src/main/resources/python/requirements.txt
redis
paho-mqtt
RPi

